test(slider): recategorize and clean up browser tests
