+++ packages/react/src/components/slider/slider.test.browser.tsx
@@ -1,6 +1,6 @@
 import { fireEvent, screen } from "@testing-library/react"
 import { vi } from "vitest"
-import { a11y, render } from "#test/browser"
+import { render } from "#test/browser"
 import { Slider } from "."
 import { noop } from "../../utils"
 
@@ -25,620 +25,6 @@ const mockRect = (el: HTMLElement, rect: Partial<DOMRect>): (() => void) => {
 }
 
 describe("<Slider />", () => {
-  test("renders component correctly", async () => {
-    await a11y(<Slider.Root defaultValue={50} />)
-  })
-
-  test("sets `displayName` correctly", () => {
-    expect(Slider.Root.displayName).toBe("SliderRoot")
-    expect(Slider.Track.displayName).toBe("SliderTrack")
-    expect(Slider.Range.displayName).toBe("SliderRange")
-    expect(Slider.Thumbs.name).toBe("SliderThumbs")
-    expect(Slider.Thumb.displayName).toBe("SliderThumb")
-    expect(Slider.Marks.name).toBe("SliderMarks")
-    expect(Slider.Mark.displayName).toBe("SliderMark")
-  })
-
-  test("sets `className` correctly", async () => {
-    await render(
-      <Slider.Root
-        data-testid="slider"
-        defaultValue={50}
-        marks={[25, 50, 75]}
-        marksProps={{ "data-testid": "marks" }}
-        trackProps={{ "data-testid": "track" }}
-      />,
-    )
-    const root = screen.getByTestId("slider")
-    const track = screen.getByTestId("track")
-    const thumb = screen.getByRole("slider")
-    const marks = screen.getAllByTestId("marks")
-    expect(root).toHaveClass("ui-slider__root")
-    expect(track).toHaveClass("ui-slider__track")
-    expect(track.children[0]).toHaveClass("ui-slider__range")
-    expect(thumb).toHaveClass("ui-slider__thumb")
-    expect(marks[0]).toHaveClass("ui-slider__mark")
-  })
-
-  test("sets aria attributes correctly", async () => {
-    await render(<Slider.Root defaultValue={50} />)
-
-    expect(screen.getByRole("slider")).toHaveAttribute("aria-valuenow", "50")
-    expect(screen.getByRole("slider")).toHaveAttribute("aria-valuetext", "50")
-    expect(screen.getByRole("slider")).toHaveAttribute("aria-valuemin", "0")
-    expect(screen.getByRole("slider")).toHaveAttribute("aria-valuemax", "100")
-    expect(screen.getByRole("slider")).toHaveAttribute(
-      "aria-orientation",
-      "horizontal",
-    )
-  })
-
-  test("disabled Slider renders correctly", async () => {
-    await render(<Slider.Root disabled />)
-
-    expect(screen.getByRole("slider")).toHaveAttribute("aria-disabled")
-  })
-
-  test("readonly Slider renders correctly", async () => {
-    await render(<Slider.Root readOnly />)
-
-    expect(screen.getByRole("slider")).toHaveAttribute("aria-readonly")
-  })
-
-  test("renders range slider with two thumbs", async () => {
-    await render(<Slider.Root defaultValue={[25, 75]} />)
-
-    const sliders = screen.getAllByRole("slider")
-    expect(sliders).toHaveLength(2)
-    expect(sliders[0]).toHaveAttribute("aria-valuenow", "25")
-    expect(sliders[1]).toHaveAttribute("aria-valuenow", "75")
-  })
-
-  test("range slider sets correct aria attributes on thumbs", async () => {
-    await render(<Slider.Root defaultValue={[20, 80]} />)
-
-    const sliders = screen.getAllByRole("slider")
-    expect(sliders[0]).toHaveAttribute("data-start")
-    expect(sliders[1]).toHaveAttribute("data-end")
-    expect(sliders[0]).toHaveAttribute("aria-valuemin", "0")
-    expect(sliders[0]).toHaveAttribute("aria-valuemax", "80")
-    expect(sliders[1]).toHaveAttribute("aria-valuemin", "20")
-    expect(sliders[1]).toHaveAttribute("aria-valuemax", "100")
-  })
-
-  test("range slider renders two hidden inputs", async () => {
-    const { container } = await render(
-      <Slider.Root name="range-slider" defaultValue={[25, 75]} />,
-    )
-
-    const inputs = container.querySelectorAll('input[type="hidden"]')
-    expect(inputs).toHaveLength(2)
-    expect(inputs[0]).toHaveAttribute("value", "25")
-    expect(inputs[1]).toHaveAttribute("value", "75")
-  })
-
-  test("single slider renders one hidden input", async () => {
-    const { container } = await render(
-      <Slider.Root name="slider" defaultValue={50} />,
-    )
-
-    const inputs = container.querySelectorAll('input[type="hidden"]')
-    expect(inputs).toHaveLength(1)
-    expect(inputs[0]).toHaveAttribute("value", "50")
-  })
-
-  test("renders marks with object format including labels", async () => {
-    await render(
-      <Slider.Root
-        defaultValue={50}
-        marks={[
-          { label: "Low", value: 25 },
-          { indicator: false, label: "High", value: 75 },
-        ]}
-      />,
-    )
-
-    expect(screen.getByText("Low")).toBeInTheDocument()
-    expect(screen.getByText("High")).toBeInTheDocument()
-  })
-
-  test("renders marks with number format", async () => {
-    await render(
-      <Slider.Root
-        data-testid="slider"
-        defaultValue={50}
-        marks={[25, 50, 75]}
-        marksProps={{ "data-testid": "mark" }}
-      />,
-    )
-
-    const marks = screen.getAllByTestId("mark")
-    expect(marks).toHaveLength(3)
-  })
-
-  test("keyboard ArrowRight increases value", async () => {
-    const onChange = vi.fn()
-
-    await render(<Slider.Root defaultValue={50} onChange={onChange} />)
-
-    const thumb = screen.getByRole("slider")
-    fireEvent.keyDown(thumb, { key: "ArrowRight" })
-
-    expect(onChange).toHaveBeenCalledWith(51)
-  })
-
-  test("keyboard ArrowLeft decreases value", async () => {
-    const onChange = vi.fn()
-
-    await render(<Slider.Root defaultValue={50} onChange={onChange} />)
-
-    const thumb = screen.getByRole("slider")
-    fireEvent.keyDown(thumb, { key: "ArrowLeft" })
-
-    expect(onChange).toHaveBeenCalledWith(49)
-  })
-
-  test("keyboard ArrowUp increases value", async () => {
-    const onChange = vi.fn()
-
-    await render(<Slider.Root defaultValue={50} onChange={onChange} />)
-
-    const thumb = screen.getByRole("slider")
-    fireEvent.keyDown(thumb, { key: "ArrowUp" })
-
-    expect(onChange).toHaveBeenCalledWith(51)
-  })
-
-  test("keyboard ArrowDown decreases value", async () => {
-    const onChange = vi.fn()
-
-    await render(<Slider.Root defaultValue={50} onChange={onChange} />)
-
-    const thumb = screen.getByRole("slider")
-    fireEvent.keyDown(thumb, { key: "ArrowDown" })
-
-    expect(onChange).toHaveBeenCalledWith(49)
-  })
-
-  test("keyboard Home sets value to min", async () => {
-    const onChange = vi.fn()
-
-    await render(<Slider.Root defaultValue={50} onChange={onChange} />)
-
-    const thumb = screen.getByRole("slider")
-    fireEvent.keyDown(thumb, { key: "Home" })
-
-    expect(onChange).toHaveBeenCalledWith(0)
-  })
-
-  test("keyboard End sets value to max", async () => {
-    const onChange = vi.fn()
-
-    await render(<Slider.Root defaultValue={50} onChange={onChange} />)
-
-    const thumb = screen.getByRole("slider")
-    fireEvent.keyDown(thumb, { key: "End" })
-
-    expect(onChange).toHaveBeenCalledWith(100)
-  })
-
-  test("keyboard PageUp increases value by ten step", async () => {
-    const onChange = vi.fn()
-
-    await render(<Slider.Root defaultValue={50} onChange={onChange} />)
-
-    const thumb = screen.getByRole("slider")
-    fireEvent.keyDown(thumb, { key: "PageUp" })
-
-    expect(onChange).toHaveBeenCalledWith(60)
-  })
-
-  test("keyboard PageDown decreases value by ten step", async () => {
-    const onChange = vi.fn()
-
-    await render(<Slider.Root defaultValue={50} onChange={onChange} />)
-
-    const thumb = screen.getByRole("slider")
-    fireEvent.keyDown(thumb, { key: "PageDown" })
-
-    expect(onChange).toHaveBeenCalledWith(40)
-  })
-
-  test("keyboard does not change value when disabled", async () => {
-    const onChange = vi.fn()
-
-    await render(<Slider.Root defaultValue={50} disabled onChange={onChange} />)
-
-    const thumb = screen.getByRole("slider")
-    fireEvent.keyDown(thumb, { key: "ArrowRight" })
-
-    expect(onChange).not.toHaveBeenCalled()
-  })
-
-  test("keyboard does not change value when readOnly", async () => {
-    const onChange = vi.fn()
-
-    await render(<Slider.Root defaultValue={50} readOnly onChange={onChange} />)
-
-    const thumb = screen.getByRole("slider")
-    fireEvent.keyDown(thumb, { key: "ArrowRight" })
-
-    expect(onChange).not.toHaveBeenCalled()
-  })
-
-  test("range slider keyboard ArrowRight on first thumb", async () => {
-    const onChange = vi.fn()
-
-    await render(<Slider.Root defaultValue={[25, 75]} onChange={onChange} />)
-
-    const sliders = screen.getAllByRole("slider")
-    fireEvent.keyDown(sliders[0]!, { key: "ArrowRight" })
-
-    expect(onChange).toHaveBeenCalledWith([26, 75])
-  })
-
-  test("range slider keyboard ArrowLeft on second thumb", async () => {
-    const onChange = vi.fn()
-
-    await render(<Slider.Root defaultValue={[25, 75]} onChange={onChange} />)
-
-    const sliders = screen.getAllByRole("slider")
-    fireEvent.keyDown(sliders[1]!, { key: "ArrowLeft" })
-
-    expect(onChange).toHaveBeenCalledWith([25, 74])
-  })
-
-  test("range slider keyboard Home on second thumb", async () => {
-    const onChange = vi.fn()
-
-    await render(<Slider.Root defaultValue={[25, 75]} onChange={onChange} />)
-
-    const sliders = screen.getAllByRole("slider")
-    fireEvent.keyDown(sliders[1]!, { key: "Home" })
-
-    expect(onChange).toHaveBeenCalledWith([25, 25])
-  })
-
-  test("range slider keyboard End on first thumb", async () => {
-    const onChange = vi.fn()
-
-    await render(<Slider.Root defaultValue={[25, 75]} onChange={onChange} />)
-
-    const sliders = screen.getAllByRole("slider")
-    fireEvent.keyDown(sliders[0]!, { key: "End" })
-
-    expect(onChange).toHaveBeenCalledWith([75, 75])
-  })
-
-  test("getAriaValueText is used for aria-valuetext", async () => {
-    const getAriaValueText = (value: number) => `${value}%`
-
-    await render(
-      <Slider.Root defaultValue={50} getAriaValueText={getAriaValueText} />,
-    )
-
-    expect(screen.getByRole("slider")).toHaveAttribute("aria-valuetext", "50%")
-  })
-
-  test("getAriaValueText is used for range slider", async () => {
-    const getAriaValueText = (value: number, index: number) =>
-      `Thumb ${index}: ${value}`
-
-    await render(
-      <Slider.Root
-        defaultValue={[25, 75]}
-        getAriaValueText={getAriaValueText}
-      />,
-    )
-
-    const sliders = screen.getAllByRole("slider")
-    expect(sliders[0]).toHaveAttribute("aria-valuetext", "Thumb 0: 25")
-    expect(sliders[1]).toHaveAttribute("aria-valuetext", "Thumb 1: 75")
-  })
-
-  test("aria-valuetext prop overrides getAriaValueText", async () => {
-    const getAriaValueText = (value: number) => `${value}%`
-
-    await render(
-      <Slider.Root
-        aria-valuetext="custom"
-        defaultValue={50}
-        getAriaValueText={getAriaValueText}
-      />,
-    )
-
-    expect(screen.getByRole("slider")).toHaveAttribute(
-      "aria-valuetext",
-      "custom",
-    )
-  })
-
-  test("respects custom min and max", async () => {
-    await render(<Slider.Root defaultValue={50} max={200} min={10} />)
-
-    const thumb = screen.getByRole("slider")
-    expect(thumb).toHaveAttribute("aria-valuemin", "10")
-    expect(thumb).toHaveAttribute("aria-valuemax", "200")
-  })
-
-  test("respects custom step", async () => {
-    const onChange = vi.fn()
-
-    await render(<Slider.Root defaultValue={50} step={5} onChange={onChange} />)
-
-    const thumb = screen.getByRole("slider")
-    fireEvent.keyDown(thumb, { key: "ArrowRight" })
-
-    expect(onChange).toHaveBeenCalledWith(55)
-  })
-
-  test("range slider with betweenThumbs prevents overlap", async () => {
-    await render(<Slider.Root betweenThumbs={10} defaultValue={[45, 55]} />)
-
-    const sliders = screen.getAllByRole("slider")
-    expect(sliders[0]).toHaveAttribute("aria-valuemax", "45")
-    expect(sliders[1]).toHaveAttribute("aria-valuemin", "55")
-  })
-
-  test("warns when max is less than min", async () => {
-    // eslint-disable-next-line @typescript-eslint/no-empty-function
-    const warnSpy = vi.spyOn(console, "warn").mockImplementation(() => {})
-
-    await render(<Slider.Root defaultValue={50} max={0} min={100} />)
-
-    expect(warnSpy).toHaveBeenCalledWith(
-      "Do not assign a number less than 'min' to 'max'",
-    )
-
-    warnSpy.mockRestore()
-  })
-
-  test("sets range CSS variables on root element", async () => {
-    await render(<Slider.Root data-testid="slider" defaultValue={50} />)
-
-    const root = screen.getByTestId("slider")
-    expect(root.style.getPropertyValue("--range-start")).toBe("0%")
-    expect(root.style.getPropertyValue("--range-end")).toBe("50%")
-  })
-
-  test("sets range CSS variables for range slider", async () => {
-    await render(<Slider.Root data-testid="slider" defaultValue={[25, 75]} />)
-
-    const root = screen.getByTestId("slider")
-    expect(root.style.getPropertyValue("--range-start")).toBe("25%")
-    expect(root.style.getPropertyValue("--range-end")).toBe("75%")
-  })
-
-  test("mark has correct data-between attribute for single slider", async () => {
-    await render(
-      <Slider.Root defaultValue={50}>
-        <Slider.Track>
-          <Slider.Range />
-          <Slider.Thumb />
-        </Slider.Track>
-        <Slider.Mark data-testid="mark-25" value={25} />
-        <Slider.Mark data-testid="mark-75" value={75} />
-      </Slider.Root>,
-    )
-
-    const mark25 = screen.getByTestId("mark-25")
-    const mark75 = screen.getByTestId("mark-75")
-
-    expect(mark25).toHaveAttribute("data-between")
-    expect(mark75).not.toHaveAttribute("data-between")
-  })
-
-  test("mark has correct data-between attribute for range slider", async () => {
-    await render(
-      <Slider.Root defaultValue={[20, 80]}>
-        <Slider.Track>
-          <Slider.Range />
-          <Slider.Thumbs />
-        </Slider.Track>
-        <Slider.Mark data-testid="mark-10" value={10} />
-        <Slider.Mark data-testid="mark-50" value={50} />
-        <Slider.Mark data-testid="mark-90" value={90} />
-      </Slider.Root>,
-    )
-
-    const mark10 = screen.getByTestId("mark-10")
-    const mark50 = screen.getByTestId("mark-50")
-    const mark90 = screen.getByTestId("mark-90")
-
-    expect(mark10).not.toHaveAttribute("data-between")
-    expect(mark50).toHaveAttribute("data-between")
-    expect(mark90).not.toHaveAttribute("data-between")
-  })
-
-  test("renders with custom children", async () => {
-    await render(
-      <Slider.Root data-testid="slider" defaultValue={50}>
-        <Slider.Track data-testid="custom-track">
-          <Slider.Range />
-          <Slider.Thumb />
-        </Slider.Track>
-      </Slider.Root>,
-    )
-
-    expect(screen.getByTestId("custom-track")).toBeInTheDocument()
-  })
-
-  test("renders range slider with custom children using SliderThumbs", async () => {
-    await render(
-      <Slider.Root data-testid="slider" defaultValue={[25, 75]}>
-        <Slider.Track>
-          <Slider.Range />
-          <Slider.Thumbs />
-        </Slider.Track>
-      </Slider.Root>,
-    )
-
-    const sliders = screen.getAllByRole("slider")
-    expect(sliders).toHaveLength(2)
-  })
-
-  test("range slider data-range attribute on range element", async () => {
-    await render(
-      <Slider.Root defaultValue={[25, 75]}>
-        <Slider.Track>
-          <Slider.Range data-testid="range" />
-          <Slider.Thumbs />
-        </Slider.Track>
-      </Slider.Root>,
-    )
-
-    const range = screen.getByTestId("range")
-    expect(range).toHaveAttribute("data-range")
-  })
-
-  test("disabled thumb has tabIndex -1", async () => {
-    await render(<Slider.Root defaultValue={50} disabled />)
-
-    const thumb = screen.getByRole("slider")
-    expect(thumb).toHaveAttribute("tabindex", "-1")
-  })
-
-  test("interactive thumb has tabIndex 0", async () => {
-    await render(<Slider.Root defaultValue={50} />)
-
-    const thumb = screen.getByRole("slider")
-    expect(thumb).toHaveAttribute("tabindex", "0")
-  })
-
-  test("renders with orientation data attribute", async () => {
-    await render(<Slider.Root data-testid="slider" defaultValue={50} />)
-
-    const root = screen.getByTestId("slider")
-    expect(root).toHaveAttribute("data-orientation", "horizontal")
-  })
-
-  test("passes inputProps to hidden input", async () => {
-    const { container } = await render(
-      <Slider.Root
-        id="test-id"
-        name="test-slider"
-        defaultValue={50}
-        required
-      />,
-    )
-
-    const input = container.querySelector('input[type="hidden"]')
-    expect(input).toHaveAttribute("name", "test-slider")
-    expect(input).toHaveAttribute("id", "test-id")
-  })
-
-  test("controlled slider value updates correctly", async () => {
-    const { rerender } = await render(<Slider.Root value={30} />)
-
-    expect(screen.getByRole("slider")).toHaveAttribute("aria-valuenow", "30")
-
-    await rerender(<Slider.Root value={70} />)
-
-    expect(screen.getByRole("slider")).toHaveAttribute("aria-valuenow", "70")
-  })
-
-  test("controlled range slider value updates correctly", async () => {
-    const { rerender } = await render(<Slider.Root value={[20, 80]} />)
-
-    const sliders = screen.getAllByRole("slider")
-    expect(sliders[0]).toHaveAttribute("aria-valuenow", "20")
-    expect(sliders[1]).toHaveAttribute("aria-valuenow", "80")
-
-    await rerender(<Slider.Root value={[30, 60]} />)
-
-    const updatedSliders = screen.getAllByRole("slider")
-    expect(updatedSliders[0]).toHaveAttribute("aria-valuenow", "30")
-    expect(updatedSliders[1]).toHaveAttribute("aria-valuenow", "60")
-  })
-
-  test("mark has correct style with --mark-position", async () => {
-    await render(
-      <Slider.Root defaultValue={50}>
-        <Slider.Track>
-          <Slider.Range />
-          <Slider.Thumb />
-        </Slider.Track>
-        <Slider.Mark data-testid="mark" value={30} />
-      </Slider.Root>,
-    )
-
-    const mark = screen.getByTestId("mark")
-    expect(mark.style.getPropertyValue("--mark-position")).toBe("30%")
-  })
-
-  test("mark has aria-hidden and role presentation", async () => {
-    await render(
-      <Slider.Root defaultValue={50}>
-        <Slider.Track>
-          <Slider.Range />
-          <Slider.Thumb />
-        </Slider.Track>
-        <Slider.Mark data-testid="mark" value={30} />
-      </Slider.Root>,
-    )
-
-    const mark = screen.getByTestId("mark")
-    expect(mark).toHaveAttribute("aria-hidden", "true")
-    expect(mark).toHaveAttribute("role", "presentation")
-  })
-
-  test("mark indicator is true by default", async () => {
-    await render(
-      <Slider.Root defaultValue={50}>
-        <Slider.Track>
-          <Slider.Range />
-          <Slider.Thumb />
-        </Slider.Track>
-        <Slider.Mark data-testid="mark" value={30} />
-      </Slider.Root>,
-    )
-
-    const mark = screen.getByTestId("mark")
-    expect(mark).toHaveAttribute("data-indicator")
-  })
-
-  test("mark indicator can be set to false", async () => {
-    await render(
-      <Slider.Root defaultValue={50}>
-        <Slider.Track>
-          <Slider.Range />
-          <Slider.Thumb />
-        </Slider.Track>
-        <Slider.Mark data-testid="mark" indicator={false} value={30} />
-      </Slider.Root>,
-    )
-
-    const mark = screen.getByTestId("mark")
-    expect(mark).not.toHaveAttribute("data-indicator")
-  })
-
-  test("renders with custom thumbProps, trackProps, rangeProps", async () => {
-    await render(
-      <Slider.Root
-        defaultValue={50}
-        rangeProps={{ "data-testid": "range" }}
-        thumbProps={{ "data-testid": "thumb" }}
-        trackProps={{ "data-testid": "track" }}
-      />,
-    )
-
-    expect(screen.getByTestId("track")).toBeInTheDocument()
-    expect(screen.getByTestId("range")).toBeInTheDocument()
-    expect(screen.getByTestId("thumb")).toBeInTheDocument()
-  })
-
-  test("range slider default children renders SliderThumbs", async () => {
-    await render(
-      <Slider.Root
-        defaultValue={[25, 75]}
-        thumbProps={{ "data-testid": "thumb" }}
-      />,
-    )
-
-    const thumbs = screen.getAllByTestId("thumb")
-    expect(thumbs).toHaveLength(2)
-  })
-
   test("pointer interaction triggers onChangeStart and onChangeEnd for single slider", async () => {
     const onChangeStart = vi.fn()
     const onChangeEnd = vi.fn()
@@ -705,7 +91,6 @@ describe("<Slider />", () => {
     const track = screen.getByTestId("track")
     const cleanup = mockRect(track, { left: 0, width: 200 })
 
-    // Click closer to the first thumb (25% = 50px)
     fireEvent.pointerDown(track, { clientX: 40, clientY: 0 })
     expect(onChangeStart).toHaveBeenCalledWith([20, 75])
 
@@ -718,7 +103,7 @@ describe("<Slider />", () => {
     cleanup()
   })
 
-  test("pointer interaction selects closest thumb for range slider (second thumb)", async () => {
+  test("pointer interaction selects the closest thumb on the second-thumb side", async () => {
     const onChangeStart = vi.fn()
     const onChangeEnd = vi.fn()
 
@@ -734,7 +119,6 @@ describe("<Slider />", () => {
     const track = screen.getByTestId("track")
     const cleanup = mockRect(track, { left: 0, width: 200 })
 
-    // Click closer to the second thumb (75% = 150px)
     fireEvent.pointerDown(track, { clientX: 160, clientY: 0 })
     expect(onChangeStart).toHaveBeenCalledWith([25, 80])
 
@@ -789,34 +173,11 @@ describe("<Slider />", () => {
     const track = screen.getByTestId("track")
     const cleanup = mockRect(track, { bottom: 200, height: 200, top: 0 })
 
-    // bottom(200) - clientY(60) = 140, 140/200 = 70% -> value ~70
     fireEvent.pointerDown(track, { clientX: 0, clientY: 60 })
     expect(onChange).toHaveBeenCalledWith(70)
 
     fireEvent.pointerUp(window, { clientX: 0, clientY: 60 })
 
     cleanup()
-  })
-
-  test("range slider keyboard PageUp on first thumb", async () => {
-    const onChange = vi.fn()
-
-    await render(<Slider.Root defaultValue={[25, 75]} onChange={onChange} />)
-
-    const sliders = screen.getAllByRole("slider")
-    fireEvent.keyDown(sliders[0]!, { key: "PageUp" })
-
-    expect(onChange).toHaveBeenCalledWith([35, 75])
-  })
-
-  test("range slider keyboard PageDown on second thumb", async () => {
-    const onChange = vi.fn()
-
-    await render(<Slider.Root defaultValue={[25, 75]} onChange={onChange} />)
-
-    const sliders = screen.getAllByRole("slider")
-    fireEvent.keyDown(sliders[1]!, { key: "PageDown" })
-
-    expect(onChange).toHaveBeenCalledWith([25, 65])
   })
 })
